First Solution Contains Duplicates
A fairly trivial problem to solve, just use a Set to store values as
you loop through the array, checking the set each iteration for a
duplicate value.
Runtime: 88ms/62.62% Memory: 45MB/49.12%
NOTE: I have moved testing of solutions to the LeetCode editor. It is a
long story but the prior path I was taking to the problems in LeetCode
produced an entirely different editor without a console.