First Solution Remove Duplicates from Sorted List
This solution is almost identical to Remove Linked List elements, the
val argument is simply replaced with the current node's value for
the values comparison. For a non-sorted list similar logic could be used
just use a Set to store all the seen values in the list.
Runtime: 72ms/98.99% Memory: 40.5MB/87.53%