First Solution Binary Tree Inorder Traversal
This is an iterative solution using a stack. It walks down the leftmost
branch, recording the nodes visited in the stack. Once a leaf is
reached, it walks back up the tree one node, recording the value and
moving on to the right leaf of the current branch. This might be
illustrated better by calling currNode root.

Recursion for inorder is more involved so iterative competes better with
for runtime. Recursion uses less memory here and there is an iterative
solution that does not use a stack so memory performance is not great.
Runtime: 72ms/73.63% Memory: 39.1MB/28.13%