First Solution Binary Tree Level Order Traversal
This solution uses a stack to hold the nodes of the current level of the
tree being evaluated and a queue to store the next level's nodes as the
current level nodes are being evaluated. This whole algorithm becomes
awkward with LeetCode's expected format for the output.
Runtime: 76ms/82.60% Memory: 40.8MB/37.19%