First Solution Symmetric Tree
This is an iterative solution that uses a helper function to evaluate
each level of the tree after the current level has been constructed. It
keeps track of each entire level in memory(O(n) memory) rather than
check each node against it's opposite and having to push the next
level's nodes in an odd order to keep track of symmetry.
Runtime: 72ms/98.90% Memory: 40.9MB/51.71%