First solution Subtract Product and Sum of Integer
Basic solution that uses the modulus operator to find to isolate each
digit of the number and apply the digit to a product and a sum variable.
An alternative approach is to create an array from the integer. This
would involve tricky type conversions or converting the integer to a
string and creating an array from that string while converting each
digit back to a number.
Runtime: 64ms/84.84% Memory: 42.5MB/24.58%