First solution Largest Perimeter Triangle
This solution sorts the array and uses Triangle Inequality Theorem 1 to
find the largest possible triangle. A way to improve memory usage may be
to sort the array highest to lowest and the use a sliding window instead
of storing each side as an individually defined variable.
Runtime: 92ms/94.96% Memory: 45.6MB/41.67%