Third Solution Find Nearest Pt w/ Same X or Y Coor
This solution uses a forEach on the points array and destructures each
nested coordinate array. Index is tracked by the second argument in the
forEach call back. I found this solution in the discussions. It once
evaluated as the fastest solution (according to the thread) but is in
the bottom third when I run it.
Runtime: 151ms/32.22% Memory: 50.4MB/72.55%