First Solution Sign of the Product of an Array
Basic solution that uses a for...of loop and just multiplies the answer
by -1 whenever a negative number is encountered.
NOTE: I am not including performance on this solution. My solution is
identical to a 48ms/39.5MB solution but mine performance takes twice as
long and takes more than 10% more memory. LeetCode is definitely pay for
ranking.