First Solution Can Make Arith. Prog. from Seq
This is a basic problem with a basic solution. Just sort the array,
record the first difference, and verify that every subsequent difference
between array values is the same.
Again not including performance metrics. My solution matches the fastest
and most memory efficient solutions but does not reflect that in the
performance measures. LeetCode is pay for performance.