First Solution Check if it is a Straight Line
This problem requires cross multiplication of the x and y differences
established by the first two points on the line with the y and x
differences between each two points on the line. This is to deal with
case of a vertical (no change in x) line. Another interesting solution
is to use the y = mx + b equation with a special case for the first two
x coordinates being 0. I will code this up at a future time.

No stats here as identical solutions provide wildly different runtimes
and memory usage. This may be due to added test cases since previous
very fast solutions.